refactor: Refactor methods in lib directory
Fixes #51

Refactor the `lib` directory to organize methods into separate files based on their functionality.

* **File Changes:**
  - Add `lib/fileUtils.go`:
    - Move methods related to file operations such as `downloadFile`, `fileExistsAndValid`, `downloadAndVerifyFile`, `findHashcatBinary`, `isExecAny`, `getCurrentHashcatVersion`, `removeExistingFile`, `writeResponseToFile`, `appendChecksumToURL`, `extractHashcatArchive`, `moveArchiveFile`, `base64ToHex`, `resourceNameOrBlank`, `downloadHashList`, and `downloadResourceFile`.
  - Add `lib/taskManager.go`:
    - Move methods related to task and attack management such as `GetNewTask`, `GetAttackParameters`, `handleAPIError`, `sendStatusUpdate`, `handleStatusUpdateError`, `handleSDKError`, `handleTaskNotFound`, `handleTaskGone`, `getZaps`, `handleGetZapsError`, `handleResponseStream`, `removeExistingZapFile`, `createAndWriteZapFile`, `SendAgentError`, and `handleSendError`.
  - Add `lib/errorUtils.go`:
    - Move methods related to error handling and validation such as `logAndSendError`, `validateHashcatDirectory`, `cleanupTempDir`, `handleAuthenticationError`, `handleConfigurationError`, `mapConfiguration`, `setNativeHashcatPath`, `getDevicesList`, `getDevices`, `extractDeviceNames`, `sendBenchmarkResults`, `createBenchmark`, `handleHeartbeatError`, `logHeartbeatSent`, and `handleStateResponse`.
  - Delete `lib/clientUtils.go` and `lib/agentClient.go` as their methods have been moved to the new files.
  - Update `lib/runners.go` and `cmd/root.go` to reflect the new file structure and update import statements and references to the moved methods.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/unclesp1d3r/CipherSwarmAgent/issues/51?shareId=XXXX-XXXX-XXXX-XXXX).
